Moved ListItem to newly created CharacterModels folder; updated using stmt in CharController
