+++ SuperHeroDB.Models/SuperHeroDB.Models.csproj
@@ -54,7 +54,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CharacterListItem.cs" />
+    <Compile Include="CharacterModels\CharacterListItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
